Issue #721 - Add support for the options real-time streaming:
Introduced an asymmetric MessagePack serdes on top of the existing JSON-
annotated message data structures:
- Incoming (majority of traffic): A direct converter that covers data types
  used by Alpaca to deserialize MessagePack data into JToken instances for
  lower processing latency.
- Outgoing (minority of traffic): A tandem converter to convert JSON messages
  from the existing pipeline into MessagePack data for ease of integration.

Added options related configuration, parameters, and utility classes/functions.

The UseMessagePack flag in the configuration is not specifiable by the end user
and is only set to true for options streaming client configuration. This is due
to the restriction from the existing WebSocket client implementation used that
the Content-Type header in the HTTP handshake request cannot be customized.